Add files via upload
the zip with raw unpacked sourcefiles.... damn, I am bad with git, and shameful for this. Tbh I wasn`t able yet to reach git manuals and as soon I have some kind of problem I normally force push, but here, I couldn`t force push as I needed the history of comits, so I pushed the vite build... but then understood it is not extracted , but minimized or so... so yeah here it is... raw and dirty, I tried.